crypto: move types::Signatures to a submodule
